Extract a couple of functions from spack.util.environment_modifications
Extracting these functions makes the module independent of
spack.util.executable
