+++ lib/spack/spack/build_systems/oneapi.py
@@ -16,8 +16,8 @@ from llnl.util.link_tree import LinkTree
 from spack.build_environment import dso_suffix
 from spack.directives import conflicts, variant
 from spack.package_base import InstallError
-from spack.util.environment import EnvironmentModifications
 from spack.util.executable import Executable
+from spack.util.sourcing import from_sourcing_file
 
 from .generic import Package
 
@@ -142,7 +142,7 @@ class IntelOneApiPackage(Package):
         # Only if environment modifications are desired (default is +envmods)
         if "~envmods" not in self.spec:
             env.extend(
-                EnvironmentModifications.from_sourcing_file(
+                from_sourcing_file(
                     self.component_prefix.env.join("vars.sh"), *self.env_script_args
                 )
             )
